Code/function/requirements.txt
pandas
matplotlib
google-cloud-storage